+++ src/backgroundblurprocessor/BackgroundFilterPaths.ts
@@ -2,19 +2,21 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /**
- * paths to assets that will be loaded from CDN
+ * Paths to assets that will be loaded from CDN
  */
 export default interface BackgroundFilterPaths {
   /**
    * Path to the web worker that will processing background segmentation.
    */
   worker: string;
+
   /**
-   * Path to the WASM file for processing background segmenation
+   * Path to the WASM file for processing background segmentation.
    */
   wasm: string;
+
   /**
-   * Path to the WASM file compiled with SIMD support for processing background segmenation
+   * Path to the WASM file compiled with SIMD support for processing background segmentation
    */
   simd: string;
 }